# [xy] Fix performance issue of fetching variables.