# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Fix performance issue of fetching variables. Avoid repeatedly calling `get_variables_dir` and read/parse the metadta.yaml.

Set the default value of MEMORY_MANAGER_VERSION to 1 since it causes the slowdown of the variables API.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Tested locally. Setting to MEMORY_MANAGER_VERSION reduces the variables API time from 27s to 1s. Reducing repeat get_variables_dir calls reduces the time from 1s to 250ms. 


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature, documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request -->